Add file Aggregator.py
This file contains an abstract class which defines the interface for all implementations that are supposed to take data samples, and compute likelihood heatmaps for visualization or for infection estimation.

There's a few implementations but the main one that now worhks is TimeSmoothAggregatorKernelDensity, the others are not finished.

Runnign the file will display a density heatmap on a randomly generated point set.